Add options to support more complex projects
Add the option to use the -i flag multiple times, this can be used
to specify source files multiple times.

Add a new flag, -p, which can be used to create more complex
include paths. This is similar to the -I flag on gcc style
compilers.

Together these options allow the bundler to be used on cpp projects
with non-flat directory structures.

Type: New Feature